+++ CHANGELOG.md
@@ -17,7 +17,8 @@ as well as to [Module version numbering](https://go.dev/doc/modules/version-numb
 ### Change
 
 - Extract the flow result reporting from `Flow.Main` to `middleware.ReportFlow`.
-  Add the middleware using `Flow.UseExecutor` to acheive a backwards compatible behavior.
+  Add the middleware using `Flow.UseExecutor` to achieve  a backwards compatible
+  behavior.
 
 ## [2.1.0](https://github.com/goyek/goyek/compare/v2.0.0...v2.1.0) - 2024-01-17